Personal Opinion Refactors
Just my personal typescript style. Instead of passing in the coords
seperately, I'm passing in a single object but then destructuring the
coords out of it. The benefit is that if you already have a `Point` you
can pass the whole thing in instead of having to reconstruct it.

The shorthand for creating an object based on the variables names as also
comes in handy here, when you already have just the x and y values.