Add some flexing to make Gameboard Adjust to Height Changes
The div that wraps the game board and the controls is now a flex container.
And the game board now knows to `flex-shrink` so that it will shrink to the
available height

There already was a nice wrapping div for the horizontal direction, so
just adding this for the vertical seemed to fix the issue!